refactor: fold npmrc-detect into auth module
Review-driven simplification:
- Merge propagateProjectNpmrcAuth into src/auth.ts since both the
  registry-url write path and the .npmrc read path are concerned with
  the same thing: .npmrc + auth env export
- Replace existsSync + readFileSync TOCTOU with try/catch on ENOENT
- Dedupe referenced env vars with a single Set instead of Set + array
- Drop detectProjectNpmrc / ProjectNpmrc from the public surface;
  test only the observable behavior of propagateProjectNpmrcAuth
- Remove narrative comment in index.ts — the function name carries it
